+++ monitoring/metrics.go
@@ -136,6 +136,7 @@ const (
 	GenerateSLAReportStr       = GenSLAReportStr + metricNameDelimiter + OPGetStr
 
 	QueryAggregatedMetricsStr = AggMetricsStr + metricNameDelimiter + OPGetStr
+	QueryThresholdCrossingStr = ThrCrossStr + metricNameDelimiter + OPGetStr
 
 	CreateTenantMetaStr = TenantMetaStr + metricNameDelimiter + OPCreateStr
 	UpdateTenantMetaStr = TenantMetaStr + metricNameDelimiter + OPUpdateStr